src/Attendance_System/requirements.txt
pipenv
psycopg2 
python-decouple
dj-database-url
djangorestframework
coverage
Faker
